lib.js log updates
- log the commands used to check react-native-cli and yarn CLI, and
  if these commands succeed
- log the options that are used to create the module
- log other steps

Note that the some of these console log messages are done as
console.info instead of verbose since they are needed to help
make sense of the console output from the third-party tools.

While the logging could use some further improvement,
it should be at least a little better than it was before.