check output of cli-command func
NOTE that a 4 second timeout is used to ensure that all of the output
is captured in the snapshot.

TBD the next step is to reduce this timeout and verify that
all of the output is still captured in the snapshot.

FUTURE TBD it is desired to rework the cli-command func to return a
Promise object so that this kind of a test timer can be removed.