remove mockable jsonfile option from lib API
(with updated tests & test snapshots)

and update documentation of the unstable fs option in the lib API

The library now uses the same fs object for both file operations and
JSON file manipulation.

Note that at this point, the library uses fs-extra by default when
using jsonfile to update package.json in the generated example.

Also note that fs-extra actually uses graceful-fs under the hood
for many if not all of the most important fs operations.