General change notes
OVERVIEW:

Windows is no longer supported for reasons discussed in #23 & #43

Moving the existing Windows C# template into `unsupported-platforms`,
will continue to be validated by the test suite for now (at least).

Removing the `uuid` package dependency which is no longer needed
(unless it is used for a supported platform again, someday).

Further discussion would be welcome *in a new issue* on GitHub.